+++ src/app/features/dashboard/dashboardCtrl.js
@@ -18,6 +18,7 @@ function (angular, $, config) {
       dynamicDashboardSrv,
       dashboardSrv,
       dashboardViewStateSrv,
+      contextSrv,
       $timeout) {
 
     $scope.editor = { index: 0 };
@@ -49,7 +50,7 @@ function (angular, $, config) {
 
         $scope.dashboard = dashboard;
         $scope.dashboardViewState = dashboardViewStateSrv.create($scope);
-        $scope.dashboardMeta = data.meta;
+        $scope.initDashboardMeta(data.meta, $scope.dashboard);
 
         dashboardKeybindings.shortcuts($scope);
 
@@ -58,6 +59,33 @@ function (angular, $, config) {
 
         $scope.appEvent("dashboard-loaded", $scope.dashboard);
       });
+    };
+
+    $scope.initDashboardMeta = function(meta, dashboard) {
+      meta.canShare = true;
+      meta.canSave = true;
+      meta.canEdit = true;
+      meta.canStar = true;
+
+      if (contextSrv.hasRole('Viewer')) {
+        meta.canSave = false;
+      }
+
+      if (meta.isHome) {
+        meta.canShare = false;
+        meta.canStar = false;
+        meta.canSave = false;
+        meta.canEdit = false;
+      }
+
+      if (dashboard.snapshot) {
+        meta.canEdit = false;
+        meta.canSave = false;
+        meta.canStar = false;
+        meta.canShare = false;
+      }
+
+      $scope.dashboardMeta = meta;
     };
 
     $scope.updateSubmenuVisibility = function() {
@@ -135,4 +163,5 @@ function (angular, $, config) {
     };
 
   });
+
 });