+++ src/app/features/panel/panelSrv.js
@@ -41,7 +41,7 @@ function (angular, _, config) {
       };
 
       $scope.updateColumnSpan = function(span) {
-        $scope.panel.span = Math.min(Math.max($scope.panel.span + span, 1), 12);
+        $scope.updatePanelSpan($scope.panel, span);
 
         $timeout(function() {
           $scope.$broadcast('render');
@@ -93,6 +93,13 @@ function (angular, _, config) {
 
       $scope.get_data = function() {
         if ($scope.otherPanelInFullscreenMode()) { return; }
+
+        if ($scope.panel.snapshotData) {
+          if ($scope.loadSnapshot) {
+            $scope.loadSnapshot($scope.panel.snapshotData);
+          }
+          return;
+        }
 
         delete $scope.panelMeta.error;
         $scope.panelMeta.loading = true;