+++ src/app/features/panel/soloPanelCtrl.js
@@ -7,16 +7,15 @@ function (angular, $) {
 
   var module = angular.module('grafana.routes');
 
-  module.controller('SoloPanelCtrl',
-    function(
-      $scope,
-      backendSrv,
-      $routeParams,
-      dashboardSrv,
-      timeSrv,
-      $location,
-      templateValuesSrv,
-      contextSrv) {
+  module.controller('SoloPanelCtrl', function(
+    $scope,
+    backendSrv,
+    $routeParams,
+    dashboardSrv,
+    timeSrv,
+    $location,
+    templateValuesSrv,
+    contextSrv) {
 
     var panelId;
 
@@ -26,12 +25,19 @@ function (angular, $) {
       var params = $location.search();
       panelId = parseInt(params.panelId);
 
-      backendSrv.getDashboard($routeParams.slug)
-        .then(function(dashboard) {
-          $scope.initPanelScope(dashboard);
-        }).then(null, function(err) {
-          $scope.appEvent('alert-error', ['Load panel error', err.message]);
-        });
+      var request;
+
+      if ($routeParams.slug) {
+        request = backendSrv.getDashboard($routeParams.slug);
+      } else {
+        request = backendSrv.get('/api/snapshots/' + $routeParams.key);
+      }
+
+      request.then(function(dashboard) {
+        $scope.initPanelScope(dashboard);
+      }).then(null, function(err) {
+        $scope.appEvent('alert-error', ['Load panel error', err.message]);
+      });
     };
 
     $scope.initPanelScope = function(dashboard) {